restructure script runner
