+++ src/renderer/scripts/editor/ScriptRunner/Compilation/runFile.ts
@@ -1,0 +1,4 @@
+import { createFileRunner } from '../runFile'
+import { ENV } from './ENV'
+
+export const runCompilationFile = createFileRunner('compile', ENV)
