+++ src/renderer/scripts/editor/ScriptRunner/Validation/ENV.ts
@@ -1,20 +1,7 @@
-import JSONTree from '../editor/JsonTree'
-import FileType from '../editor/FileType'
-import TabSystem from '../TabSystem'
-import LightningCache from '../editor/LightningCache'
-
-export function prepareRun(code: string) {
-	try {
-		return function(Bridge: unknown) {
-			return eval(code)
-		}
-	} catch (err) {
-		throw err
-	}
-}
-export function run(code: string, env: unknown) {
-	return prepareRun(code)(env)
-}
+import JSONTree from '../../JsonTree'
+import FileType from '../../FileType'
+import TabSystem from '../../../TabSystem'
+import LightningCache from '../../LightningCache'
 
 export const ENV = (
 	Node: JSONTree,
