+++ src/renderer/scripts/editor/ScriptRunner/Validation/runFile.ts
@@ -1,0 +1,4 @@
+import { createFileRunner } from '../runFile'
+import { ENV } from './ENV'
+
+export const runValidationFile = createFileRunner('validate', ENV)
