+++ src/renderer/scripts/editor/ScriptRunner/run.ts
@@ -1,0 +1,12 @@
+export function prepareRun(code: string) {
+	try {
+		return function(Bridge: unknown) {
+			return eval(code)
+		}
+	} catch (err) {
+		throw err
+	}
+}
+export function run(code: string, env: unknown) {
+	return prepareRun(code)(env)
+}
